pyproject.toml
[project]
name = "youtube-toolkit"
version = "0.1.0"
description = "CLI for YouTube channel analysis and transcript downloading"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
authors = [{ name = "Manuel Lorente", email = "info@mlorente.dev" }]
keywords = ["youtube", "analytics", "cli", "transcripts", "engagement-metrics"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Video",
]
dependencies = [
    "google-api-python-client>=2.112.0",
    "google-auth-oauthlib>=1.2.0",
    "youtube-transcript-api>=1.2.2",
    "pyyaml>=6.0",
    "typer>=0.15.0",
    "rich>=13.0",
    "pydantic-settings>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.8.0",
    "mypy>=1.13",
    "types-PyYAML>=6.0",
]

[project.scripts]
youtube-toolkit = "src.main:app"

[project.urls]
Homepage = "https://github.com/mlorentedev/youtube-toolkit"
Documentation = "https://mlorentedev.github.io/youtube-toolkit/"
Repository = "https://github.com/mlorentedev/youtube-toolkit"
Issues = "https://github.com/mlorentedev/youtube-toolkit/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "A", "SIM", "TCH"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = ["googleapiclient", "googleapiclient.*", "youtube_transcript_api", "youtube_transcript_api.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=src --cov-report=term-missing"
